poc: Move test vector generation out of unit tests
Remove the TEST_VECTOR and TEST_VECTOR_PATH environment variables and
move test vector generation logic into a new tool, `gen_test_vec.py`.

The test vectors are complete with the following exceptions:

1. We currently don't have test vectors for Prio3MultihotCountVec. This
   VDAF is new as of draft 10, but we neglected to add test vectors for
   it. This change checks in the test vectors.

2. The test vector for IdpfBBCGGI21 was regenerated after this IDPF was
   renamed. Also, as of 01291c2c7c7568f8d31d01aa4d3e9df653d6b447, we are
   generating this with a proper nonce rather than an adhoc binder. To
   confirm that nothing has changed, you can modify
   `gen_test_vec_for_idpf()` by replacing Lines

    ```
    nonce = gen_rand(idpf.NONCE_SIZE)
    ```

    with

    ```
    nonce = bytes([0x73,0x6f,0x6d,0x65,0x20,0x6e,0x6f,0x6e,0x63,0x65])
    ```

Also, this change opts in `gen_test_vec.py` and
`plot_prio3_multiproof_robustness.py` into linting and type checking.
