+++ poc/README.md
@@ -38,17 +38,15 @@ sage -python -m unittest
 
 ## Generating test vectors
 
-To generate test vectors, set environment variable `TEST_VECTOR` to
-be `TRUE` when running tests:
+To generate test vectors, run:
 
 ```
-TEST_VECTOR=TRUE sage -python -m unittest --quiet
+sage -python gen_test_vec.py
 ```
 
 Users can also specify a custom path to generate the test vectors in
-environment variable `TEST_VECTOR_PATH`. For example, to generate
-test vectors for Prio3 VDAFs into path `test_vec/00`:
+environment variable `TEST_VECTOR_PATH`:
 
 ```
-TEST_VECTOR=TRUE TEST_VECTOR_PATH=test_vec/00 sage -python -m unittest --quiet tests/test_vdaf_prio3.py
+TEST_VECTOR_PATH=path/to/test_vec sage -python gen_test_vec.py
 ```
