+++ poc/tests/idpf_util.py
@@ -1,9 +1,7 @@
-import json
-import os
 from functools import reduce
 from typing import Sequence
 
-from vdaf_poc.common import TEST_VECTOR_PATH, gen_rand, vec_add
+from vdaf_poc.common import gen_rand, vec_add
 from vdaf_poc.idpf import Idpf
 
 
@@ -81,35 +79,3 @@ def test_idpf_exhaustive(idpf: Idpf, alpha: int) -> None:
             if got != want:
                 print('error: {0:b} {1:b} {2}: got {3}; want {4}'.format(
                     alpha, prefix, level, got, want))
-
-
-def gen_test_vec(idpf: Idpf, alpha: int, test_vec_instance: int) -> None:
-    beta_inner = []
-    for level in range(idpf.BITS - 1):
-        beta_inner.append([idpf.field_inner(level)] * idpf.VALUE_LEN)
-    beta_leaf = [idpf.field_leaf(idpf.BITS - 1)] * idpf.VALUE_LEN
-    rand = gen_rand(idpf.RAND_SIZE)
-    nonce = gen_rand(idpf.NONCE_SIZE)
-    (public_share, keys) = idpf.gen(alpha, beta_inner, beta_leaf, nonce, rand)
-
-    printable_beta_inner = [
-        [str(elem.as_unsigned()) for elem in value] for value in beta_inner
-    ]
-    printable_beta_leaf = [str(elem.as_unsigned()) for elem in beta_leaf]
-    printable_keys = [key.hex() for key in keys]
-    test_vec = {
-        'bits': int(idpf.BITS),
-        'alpha': str(alpha),
-        'beta_inner': printable_beta_inner,
-        'beta_leaf': printable_beta_leaf,
-        'nonce': nonce.hex(),
-        'public_share': public_share.hex(),
-        'keys': printable_keys,
-    }
-
-    os.system('mkdir -p {}'.format(TEST_VECTOR_PATH))
-    filename = '{}/{}_{}.json'.format(TEST_VECTOR_PATH, idpf.test_vec_name,
-                                      test_vec_instance)
-    with open(filename, 'w') as f:
-        json.dump(test_vec, f, indent=4, sort_keys=True)
-        f.write('\n')
