+++ poc/tests/test_idpf_bbcggi21.py
@@ -1,16 +1,14 @@
 import unittest
 from typing import cast
 
-from tests.idpf_util import gen_test_vec, test_idpf, test_idpf_exhaustive
-from vdaf_poc.common import TEST_VECTOR, from_be_bytes, gen_rand, vec_add
+from tests.idpf_util import test_idpf, test_idpf_exhaustive
+from vdaf_poc.common import from_be_bytes, gen_rand, vec_add
 from vdaf_poc.field import Field
 from vdaf_poc.idpf_bbcggi21 import IdpfBBCGGI21
 
 
 class TestIdpfBBCGGI21(unittest.TestCase):
     def test(self) -> None:
-        if TEST_VECTOR:
-            gen_test_vec(IdpfBBCGGI21(2, 10), 0, 0)
         test_idpf(
             IdpfBBCGGI21(2, 16),
             0b1111000011110000,
