+++ poc/vdaf_poc/vdaf.py
@@ -1,12 +1,9 @@
 """Definition of VDAFs."""
 
-import json
-import os
 from abc import ABCMeta, abstractmethod
-from typing import Any, Generic, Optional, TypedDict, TypeVar, Union, cast
+from typing import Any, Generic, TypeVar, Union
 
-from vdaf_poc.common import (TEST_VECTOR_PATH, format_dst, gen_rand,
-                             print_wrapped_line, to_le_bytes)
+from vdaf_poc.common import format_dst, gen_rand
 from vdaf_poc.field import Field
 
 Measurement = TypeVar("Measurement")
@@ -19,6 +16,7 @@ AggResult = TypeVar("AggResult")
 PrepState = TypeVar("PrepState")
 PrepShare = TypeVar("PrepShare")
 PrepMessage = TypeVar("PrepMessage")
+F = TypeVar("F", bound=Field)
 
 
 class Vdaf(
@@ -209,29 +207,6 @@ class Vdaf(
         pass
 
 
-class PrepTestVectorDict(Generic[Measurement], TypedDict):
-    measurement: Measurement
-    nonce: str
-    input_shares: list[str]
-    prep_shares: list[list[str]]
-    prep_messages: list[str]
-    out_shares: list[list[str]]
-    rand: str
-    public_share: str
-
-
-class TestVectorDict(Generic[Measurement, AggParam, AggResult], TypedDict):
-    shares: int
-    verify_key: str
-    agg_param: AggParam
-    prep: list[PrepTestVectorDict[Measurement]]
-    agg_shares: list[str]
-    agg_result: Optional[AggResult]
-
-
-F = TypeVar("F", bound=Field)
-
-
 # NOTE: This function is excerpted in the document, as the figure
 # {{run-vdaf}}. Its width should be limited to 69 columns to avoid
 # warnings from xml2rfc.
@@ -252,9 +227,7 @@ def run_vdaf(
         verify_key: bytes,
         agg_param: AggParam,
         nonces: list[bytes],
-        measurements: list[Measurement],
-        print_test_vec: bool = False,
-        test_vec_instance: int = 0) -> AggResult:
+        measurements: list[Measurement]) -> AggResult:
     """
     Run the VDAF on a list of measurements.
 
@@ -274,19 +247,6 @@ def run_vdaf(
             "measurements and nonces lists have different lengths"
         )
 
-    # REMOVE ME
-    test_vec: TestVectorDict[Measurement, AggParam, AggResult] = {
-        'shares': vdaf.SHARES,
-        'verify_key': verify_key.hex(),
-        'agg_param': agg_param,
-        'prep': [],
-        'agg_shares': [],
-        'agg_result': None,  # set below
-    }
-    type_params = vdaf.test_vec_set_type_param(
-        cast(dict[str, Any], test_vec)
-    )
-
     out_shares = []
     for (nonce, measurement) in zip(nonces, measurements):
         assert len(nonce) == vdaf.NONCE_SIZE
@@ -295,23 +255,6 @@ def run_vdaf(
         rand = gen_rand(vdaf.RAND_SIZE)
         (public_share, input_shares) = \
             vdaf.shard(measurement, nonce, rand)
-
-        # REMOVE ME
-        prep_test_vec: PrepTestVectorDict[Measurement] = {
-            'measurement': measurement,
-            'nonce': nonce.hex(),
-            'input_shares': [],
-            'prep_shares': [[] for _ in range(vdaf.ROUNDS)],
-            'prep_messages': [],
-            'out_shares': [],
-            'rand': rand.hex(),
-            'public_share': vdaf.test_vec_encode_public_share(
-                public_share
-            ).hex()
-        }
-        for input_share in input_shares:
-            prep_test_vec['input_shares'].append(
-                vdaf.test_vec_encode_input_share(input_share).hex())
 
         # Each Aggregator initializes its preparation state.
         prep_states = []
@@ -325,18 +268,10 @@ def run_vdaf(
             prep_states.append(state)
             outbound_prep_shares.append(share)
 
-        # REMOVE ME
-        for prep_share in outbound_prep_shares:
-            prep_test_vec['prep_shares'][0].append(
-                vdaf.test_vec_encode_prep_share(prep_share).hex())
-
         # Aggregators recover their output shares.
         for i in range(vdaf.ROUNDS - 1):
             prep_msg = vdaf.prep_shares_to_prep(agg_param,
                                                 outbound_prep_shares)
-            # REMOVE ME
-            prep_test_vec['prep_messages'].append(
-                vdaf.test_vec_encode_prep_msg(prep_msg).hex())
 
             outbound_prep_shares = []
             for j in range(vdaf.SHARES):
@@ -344,33 +279,17 @@ def run_vdaf(
                 assert isinstance(out, tuple)
                 (prep_states[j], prep_share) = out
                 outbound_prep_shares.append(prep_share)
-            # REMOVE ME
-            for prep_share in outbound_prep_shares:
-                prep_test_vec['prep_shares'][i+1].append(
-                    vdaf.test_vec_encode_prep_share(prep_share).hex()
-                )
 
         # The final outputs of the prepare phase are the output
         # shares.
         prep_msg = vdaf.prep_shares_to_prep(agg_param,
                                             outbound_prep_shares)
-        # REMOVE ME
-        prep_test_vec['prep_messages'].append(
-            vdaf.test_vec_encode_prep_msg(prep_msg).hex())
 
         outbound_out_shares = []
         for j in range(vdaf.SHARES):
             out_share = vdaf.prep_next(prep_states[j], prep_msg)
             assert not isinstance(out_share, tuple)
             outbound_out_shares.append(out_share)
-
-        # REMOVE ME
-        for out_share in outbound_out_shares:
-            prep_test_vec['out_shares'].append([
-                to_le_bytes(x.as_unsigned(), x.ENCODED_SIZE).hex()
-                for x in out_share
-            ])
-        test_vec['prep'].append(prep_test_vec)
 
         out_shares.append(outbound_out_shares)
 
@@ -382,82 +301,12 @@ def run_vdaf(
         out_shares_j = [out[j] for out in out_shares]
         agg_share_j = vdaf.aggregate(agg_param, out_shares_j)
         agg_shares.append(agg_share_j)
-        # REMOVE ME
-        test_vec['agg_shares'].append(
-            vdaf.test_vec_encode_agg_share(agg_share_j).hex())
 
     # Collector unshards the aggregate.
     num_measurements = len(measurements)
     agg_result = vdaf.unshard(agg_param, agg_shares,
                               num_measurements)
-    # REMOVE ME
-    test_vec['agg_result'] = agg_result
-    if print_test_vec:
-        pretty_print_vdaf_test_vec(vdaf, test_vec, type_params)
-
-        os.system('mkdir -p {}'.format(TEST_VECTOR_PATH))
-        filename = '{}/{}_{}.json'.format(
-            TEST_VECTOR_PATH,
-            vdaf.test_vec_name,
-            test_vec_instance,
-        )
-        with open(filename, 'w', encoding="UTF-8") as f:
-            json.dump(test_vec, f, indent=4, sort_keys=True)
-            f.write('\n')
-
     return agg_result
-
-
-def pretty_print_vdaf_test_vec(
-        vdaf: Vdaf[
-            Measurement, AggParam, Any, Any, Any, Any, AggResult, Any, Any, Any
-        ],
-        typed_test_vec: TestVectorDict[Measurement, AggParam, AggResult],
-        type_params: list[str]) -> None:
-    test_vec = cast(dict[str, Any], typed_test_vec)
-    print('---------- {} ---------------'.format(vdaf.test_vec_name))
-    for type_param in type_params:
-        print('{}: {}'.format(type_param, test_vec[type_param]))
-    print('verify_key: "{}"'.format(test_vec['verify_key']))
-    if test_vec['agg_param'] is not None:
-        print('agg_param: {}'.format(test_vec['agg_param']))
-
-    for (n, prep_test_vec) in enumerate(test_vec['prep']):
-        print('upload_{}:'.format(n))
-        print('  measurement: {}'.format(prep_test_vec['measurement']))
-        print('  nonce: "{}"'.format(prep_test_vec['nonce']))
-        print('  public_share: >-')
-        print_wrapped_line(prep_test_vec['public_share'], tab=4)
-
-        # Shard
-        for (i, input_share) in enumerate(prep_test_vec['input_shares']):
-            print('  input_share_{}: >-'.format(i))
-            print_wrapped_line(input_share, tab=4)
-
-        # Prepare
-        for (i, (prep_shares, prep_msg)) in enumerate(
-                zip(prep_test_vec['prep_shares'],
-                    prep_test_vec['prep_messages'])):
-            print('  round_{}:'.format(i))
-            for (j, prep_share) in enumerate(prep_shares):
-                print('    prep_share_{}: >-'.format(j))
-                print_wrapped_line(prep_share, tab=6)
-            print('    prep_message: >-')
-            print_wrapped_line(prep_msg, tab=6)
-
-        for (j, out_shares) in enumerate(prep_test_vec['out_shares']):
-            print('  out_share_{}:'.format(j))
-            for out_share in out_shares:
-                print('    - {}'.format(out_share))
-
-    # Aggregate
-    for (j, agg_share) in enumerate(test_vec['agg_shares']):
-        print('agg_share_{}: >-'.format(j))
-        print_wrapped_line(agg_share, tab=2)
-
-    # Unshard
-    print('agg_result: {}'.format(test_vec['agg_result']))
-    print()
 
 
 def test_vdaf(
@@ -475,9 +324,7 @@ def test_vdaf(
         ],
         agg_param: AggParam,
         measurements: list[Measurement],
-        expected_agg_result: AggResult,
-        print_test_vec: bool = False,
-        test_vec_instance: int = 0) -> None:
+        expected_agg_result: AggResult) -> None:
     # Test that the algorithm identifier is in the correct range.
     assert 0 <= vdaf.ID and vdaf.ID < 2 ** 32
 
@@ -488,9 +335,7 @@ def test_vdaf(
                           verify_key,
                           agg_param,
                           nonces,
-                          measurements,
-                          print_test_vec,
-                          test_vec_instance)
+                          measurements)
     if agg_result != expected_agg_result:
         print('vdaf test failed ({} on {}): unexpected result: got {}; want {}'
               .format(vdaf.test_vec_name, measurements, agg_result,
