+++ test_vec/08/IdpfBBCGGI21_0.json
@@ -1,0 +1,52 @@
+{
+    "alpha": "0",
+    "beta_inner": [
+        [
+            "0",
+            "0"
+        ],
+        [
+            "1",
+            "1"
+        ],
+        [
+            "2",
+            "2"
+        ],
+        [
+            "3",
+            "3"
+        ],
+        [
+            "4",
+            "4"
+        ],
+        [
+            "5",
+            "5"
+        ],
+        [
+            "6",
+            "6"
+        ],
+        [
+            "7",
+            "7"
+        ],
+        [
+            "8",
+            "8"
+        ]
+    ],
+    "beta_leaf": [
+        "9",
+        "9"
+    ],
+    "bits": 10,
+    "keys": [
+        "000102030405060708090a0b0c0d0e0f",
+        "101112131415161718191a1b1c1d1e1f"
+    ],
+    "nonce": "000102030405060708090a0b0c0d0e0f",
+    "public_share": "54c50d0a193b2b5d22d98b69cacff166bad7c487acc275da62c9494970c8c530691071489a38f73f182c52e1c6581a727a18bbc9dd4eefd9b84dfa08174ae87e6513ca16b15e42fe6bb43bb883c58260a9dc907478cec4bc7e5351049a376a612b03c75446e5941d0bdf4c1a105f1b8365f54a7712051173961d6a6826303ed083ee3e1aeee2db4a9f42a53c4f5c3c05f9909f905e7908c5b7068927912d4319c391d284006c3376c0f1a06307d532000b6b9740ce776f24e0413c8745c9f79d63845fd03f68bd07d0639771680453ea261eba1f5c44f23e702cb3002e98ddd7e36a01f245401c5d7762c85768c23bf06dc6b6b32d482a093504a3d507cda0b8b5e3ff36562cb8263da01ce5e9f801463a7115eb436c15867ab4b64f01400c8c7aabe7868e5e8152a2ee30cd9d7625fb6c261cd2deafe75f81f25701398398d6d2021ce160a9652b9f02223d8ca8af4c45f7023bdceb56b2b1e407cf6ef0f485e0291ea082bc632938f75bd2cf2b4441e8b02f"
+}
